Spit queue adapter core
